Move the missed-messages delays into constants.py
