+++ app/backend/src/couchers/constants.py
@@ -99,6 +99,11 @@ ACTIVENESS_PROBE_TIME_REMINDERS = [timedelta(days=0), timedelta(days=2, hours=8)
 # total time from initiation after which to expire the probe
 ACTIVENESS_PROBE_EXPIRY_TIME = timedelta(days=4)
 
+# how long a message must go unseen before we email the user about it
+MISSED_MESSAGES_DELAY = timedelta(minutes=5)
+# ... unless we could reach them by push, in which case they've already been told about it once
+MISSED_MESSAGES_DELAY_WITH_PUSH = timedelta(hours=24)
+
 HOST_REQUEST_MAX_REMINDERS = 1
 HOST_REQUEST_REMINDER_INTERVAL = timedelta(days=2)
